pyproject.toml
[tool.poetry]
name = "kaigen-automate"
version = "0.1.0"
description = "Automation tool for downloading documents from Kaigen and uploading to Google Drive"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.40.0"
google-api-python-client = "^2.100.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
python-dotenv = "^1.0.0"
tenacity = "^8.2.0"
pydantic = "^2.5.0"
google-generativeai = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.12.0"
black = "^23.11.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

